modifica ed eliminazione di un'asta solo da proprietari

+++ app/auctions.js
@@ -33,7 +33,7 @@ router.post('/auctions', verifyToken, async (req, res) => {
       let id = last_auction?.id !== undefined ? last_auction.id: -1;
       id++;
       
-      const new_auction = {id, titolo_asta, desc_asta, scadenza, offerta_iniziale};
+      const new_auction = {id, proprietario: req.userId, titolo_asta, desc_asta, scadenza, offerta_iniziale};
       await mongo.collection('auctions').insertOne(new_auction);
       res.send("Nuova asta aggiunta con successo!")
     }
@@ -70,30 +70,38 @@ router.get('/auctions/:id', async (req, res) => {
   }
 });
 
-router.put('/auctions/:id', async (req, res) => {
+router.put('/auctions/:id', verifyToken, async (req, res) => {
   try {
     const mongo = await db.connect2db();
 
     const id = parseInt(req.params.id);
+    if (isNaN(id)) {
+      console.error("Invalid ID:", req.params.id);
+      return res.status(400).send("Invalid ID");
+    }
     const titolo_asta = req.body.titolo_asta;
     const desc_asta = req.body.desc_asta;
-    const scadenza = req.body.scadenza;
-    const offerta_iniziale = req.body.offerta_iniziale;  
 
-    const auction = {id, titolo_asta, desc_asta, scadenza, offerta_iniziale};
+    const updated_auction = {titolo_asta, desc_asta};
 
     const query = {id: id};
-    console.log(query)
 
-    await mongo.collection("auctions").updateOne(query, {$set: auction});
-    res.send(auction);
+    const auction = await mongo.collection("auctions").findOne(query);
+    
+    if (auction.proprietario !== req.userId) {
+      console.error('UPSIE')
+      return res.status(405).send('Utente non autorizzato alla modifica!')
+    }
+
+    await mongo.collection("auctions").updateOne(query, {$set: updated_auction});
+    res.send("Auction updated successfully");
   } catch (error) {
     console.error("Errore:", error);
     res.status(500).send("Errore interno del server")
   }
 });
 
-router.delete('/auctions/:id', async (req, res) => {
+router.delete('/auctions/:id', verifyToken, async (req, res) => {
   try {
     const mongo = await db.connect2db();
     console.log("Connesso al database");
@@ -105,15 +113,16 @@ router.delete('/auctions/:id', async (req, res) => {
     }
     const query = {id: id};
 
-    console.log("id:", id)
-    console.log("query:", query)
+    const auction = await mongo.collection("auctions").findOne(query);
+    console.log(auction)
 
-    const result = await mongo.collection("auctions").deleteOne(query);
-    if (result.deletedCount === 0) {
-      console.log("No document found with the given ID");
-      return res.status(404).send("Auction not found");
+    if (auction.proprietario !== req.userId) {
+      console.error('ERRORE')
+      return res.status(405).send('Utente non autorizzato alla cancellazione!')
     }
-    res.status().send("Auction deleted successfully")
+
+    await mongo.collection("auctions").deleteOne(query);
+    res.send("Auction deleted successfully")
   } catch (error) {
     console.error("Errore:", error);
     res.status(500).send("Errore interno del server") 
@@ -121,3 +130,5 @@ router.delete('/auctions/:id', async (req, res) => {
 });
 
 module.exports = router;
+
+// curl -X DELETE http://localhost:3000/api/auctions/3 --cookie "token=tuo_token_valido"
